Move getNextTabbableElement into TabHandler class
